1、设备管理中部门为必填项。 2、修改了每个实体的ID的生成策略，由之前的自动生成改为通过序列生成。 3、修改了在文件检索时，选择部门查询时报别名重名的问题。 4、修改了选择部门时的界面问题。

+++ src/main/java/com/njmd/zfms/web/controller/FileTypeMgrController.java
@@ -45,7 +45,7 @@ public class FileTypeMgrController extends BaseController
 		// 设置默认排序方式
 		if (!page.isOrderBySetted())
 		{
-			page.setOrder(Page.DESC);
+			page.setOrder(Page.ASC);	
 			page.setOrderBy("typeId");
 		}
 		List<PropertyFilter> filters = HibernateWebUtils.buildPropertyFilters(request);
@@ -88,18 +88,24 @@ public class FileTypeMgrController extends BaseController
 	}
 
 	/** 删除 */
+	//Edit by 孙强伟，增加系统默认设置无法删除提示
 	@RequestMapping(value = "/delete/{id}")
 	@ResponseBody
 	public ResultInfo delete(HttpServletRequest request, Model model, @PathVariable("id") Long id) throws Exception
 	{
-		int resultTag = fileTypeInfoService.delete(id);
-		if (resultTag == ResultConstants.DELETE_SUCCEED)
-		{
-			return ResultInfo.saveMessage(ResultConstants.getResultInfo(resultTag, INFORMATION_PARAMAS), REDIRECT_PATH);
-		}
-		else
-		{
-			return ResultInfo.saveErrorMessage(ResultConstants.getResultInfo(resultTag, INFORMATION_PARAMAS));
+		//判断是否为默认记录
+		if(0!=id){
+			int resultTag = fileTypeInfoService.delete(id);
+			if (resultTag == ResultConstants.DELETE_SUCCEED)
+			{
+				return ResultInfo.saveMessage(ResultConstants.getResultInfo(resultTag, INFORMATION_PARAMAS), REDIRECT_PATH);
+			}
+			else
+			{
+				return ResultInfo.saveErrorMessage(ResultConstants.getResultInfo(resultTag, INFORMATION_PARAMAS));
+			}
+		}else{
+			return ResultInfo.saveErrorMessage(ResultConstants.getResultInfo(ResultConstants.DELETE_SYSTEM_DEFAULT_ERROR));
 		}
 	}
